fix(package.json): use peerDependencies because of incompatibilities between npm 2 - 3
If we define the dependencies as normal dependencies, those will be installed differently based on
the npm version. With npm@3 the dependencies are flattened on the top level and they are available
to all libraries. With npm@2 those dependencies will be only available to this package, making it
impossible to use eslint properly.

Refs #6